requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
pywin32>=228 